requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
django-allauth==0.32.0
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-datetime-widget==0.9.3
django-environ==0.4.3
django-imagekit==4.0.1
django-model-utils==3.0.0
django-summernote==0.8.7
django-taggit==0.22.1
idna==2.5
oauthlib==2.0.2
olefile==0.44
pilkit==2.0
Pillow==4.1.1
psycopg2==2.7.1
python-openid==2.2.5
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
